* Win64 port: Patches for Contrib Plugins.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@9034 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/keybinder/keybinder.cpp
@@ -2283,7 +2283,7 @@ void wxKeyConfigPanel::AddRootIfMissing(const wxString &rootname)
 wxCmd *wxKeyConfigPanel::GetSelCmd() const
 // ----------------------------------------------------------------------------
 {
-	int id;
+	intptr_t id;
 
 	if (IsUsingTreeCtrl()) {
 
@@ -2301,7 +2301,7 @@ wxCmd *wxKeyConfigPanel::GetSelCmd() const
 		if (sel < 0)
 			return NULL;
 
-		id = (long)m_pCommandsList->GetClientData(sel);
+		id = (intptr_t)m_pCommandsList->GetClientData(sel);
 	}
 
 	return m_kBinder.GetCmd(id);
